fix(map): use new translation files
using just the language code resulted in multiple translation
sources (e.g. portuguese and brazilian portuguese) trampling on one
another, seemingly reverting translations.

This PR updates the localization to use the country-specific
translations, and in the case of portuguese, combines the 2
translations (given we have more for brazillian), prioritizing the
non-brazillian.