chore(web-client): remove model classes and class-validator
With the exception of default values,
the functionality of these classes (i.e. data validation) is not
being used at all in the web-client code, so we can just use the
model interfaces and data objects directly.

For default values, I've introduced the concept of "init"
functions that allow some properties to be made optional,
which means it will now be explicit throughout the
codebase which objects are fully-initialized documents,
and which are partial.