feat: added partial protection route guard, used in find, new request
the goal of this route guard is to check whether the user authenticated themselves and have completed the onboarding procedure of verifying the phone number and the default address. If the user has completed the onboarding procedure a flag is stored in the redux auth state indicating that the user has done so. If the onboarding procedure hasn’t been completed, the guard doesn’t redirect the user. This way, the partial protection guard is different from the protection guard in that it only checks and stores a flag rather than restricting the access